+++ src/main/java/alex/home/angular/dto/ProductField.java
@@ -4,19 +4,19 @@ import java.io.Serializable;
 
 public class ProductField implements Serializable {
     
-    public Long productId;
+    public Integer productId;
     public String columnName;
     public String value;
     
     public ProductField() {}
     
-    public ProductField(Long productId, String columnName, String value) {
+    public ProductField(Integer productId, String columnName, String value) {
         this.productId = productId;
         this.columnName = columnName;
         this.value = value;
     }
 
-    public Long getProductId() {
+    public Integer getProductId() {
         return productId;
     }
